+++ src/util/meson.build
@@ -1,8 +1,4 @@
-util_src = files([
-  'util_env.cpp',
-  'util_string.cpp',
-  'util_gdi.cpp',
-  'util_luid.cpp',
+util_src = [
   'util_matrix.cpp',
   
   'com/com_guid.cpp',
@@ -15,7 +11,22 @@ util_src = files([
   
   'sha1/sha1.c',
   'sha1/sha1_util.cpp',
-])
+]
+
+util_src_win32 = [
+  'util_gdi.cpp',
+  'com/com_guid_values.cpp',
+
+  'platform/util_luid_win32.cpp',
+  'platform/util_env_win32.cpp',
+  'platform/util_string_win32.cpp'
+]
+
+if dxvk_platform == 'windows'
+  util_src += util_src_win32
+else
+  error('Unknown platform for util')
+endif
 
 util_lib = static_library('util', util_src,
   include_directories : [ dxvk_include_path ],
